add warning on asset for completeness
warning added: Asset attributes must have a variation generated for the selected channel (and locale if your reference file is localisable) to have their completeness filled. If the yellow dot remains even if you have an asset file associated to your product, check if it has a variation for your channel.