Fix the scan by Algolia
As the dropdown is in JS, Algolia cannot scan the links to follow. The fix is just to generate it from the versions.json. As it follows the links from the serenity version at first, the links will always be up to date. Indeed, only the versions.json from master is the source of truth about the dropdown.