Update category ACLs for EC Early Access
As we will soon deliver EC Early Access in SaaS only (not v7), I need to update and add new specific ACLs to the article named "Manage the interface and actions accesses" in the  Permissions on categories section. 
The merge will be only done after Master production branch is unfrozen and feature flag activated to release the feature.